Add useHazardDataFetcher and refactored search-bar. Wrap getHazardData in useCallback.
